[MOVED] 클린아키텍쳐 적용
