+++ src/main/kotlin/com/dotori/v2/domain/email/domain/repository/EmailCertificateRepository.kt
@@ -1,0 +1,10 @@
+package com.dotori.v2.domain.email.domain.repository
+
+import com.dotori.v2.domain.email.domain.entity.EmailCertificate
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface EmailCertificateRepository: JpaRepository<EmailCertificate, Long> {
+    fun deleteByEmail(email: String)
+
+    fun findByKey(key: String): EmailCertificate?
+}
