+++ src/main/kotlin/com/dotori/v2/domain/member/domain/repository/MemberRepository.kt
@@ -1,0 +1,9 @@
+package com.dotori.v2.domain.member.domain.repository
+
+import com.dotori.v2.domain.member.domain.entity.Member
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface MemberRepository: JpaRepository<Member, Long> {
+    fun findByEmail(email: String?): Member?
+    fun existsByEmail(email: String): Boolean
+}
