+++ src/main/kotlin/com/dotori/v2/global/annotation/ReadOnlyUseCase.kt
@@ -1,0 +1,8 @@
+package com.dotori.v2.global.annotation
+
+import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
+
+@Service
+@Transactional(rollbackFor = [Exception::class], readOnly = true)
+annotation class ReadOnlyUseCase()
