+++ src/main/kotlin/com/dotori/v2/global/config/security/handler/CustomAccessDeniedHandler.kt
@@ -1,0 +1,22 @@
+package com.dotori.v2.global.config.security.handler
+
+import org.slf4j.LoggerFactory
+import org.springframework.security.access.AccessDeniedException
+import org.springframework.security.web.access.AccessDeniedHandler
+import org.springframework.stereotype.Component
+import java.io.IOException
+import javax.servlet.ServletException
+import javax.servlet.http.HttpServletRequest
+import javax.servlet.http.HttpServletResponse
+
+@Component
+class CustomAccessDeniedHandler : AccessDeniedHandler {
+
+    private val log = LoggerFactory.getLogger(this::class.simpleName)
+
+    @Throws(IOException::class, ServletException::class)
+    override fun handle(request: HttpServletRequest, response: HttpServletResponse, accessDeniedException: AccessDeniedException) {
+        log.info("=== Access Denied ===")
+        response.sendError(HttpServletResponse.SC_FORBIDDEN)
+    }
+}
