+++ src/main/kotlin/com/dotori/v2/global/config/security/jwt/JwtReqFilter.kt
@@ -1,0 +1,26 @@
+package com.dotori.v2.global.config.security.jwt
+
+import org.springframework.security.core.Authentication
+import org.springframework.security.core.context.SecurityContextHolder
+import org.springframework.stereotype.Component
+import org.springframework.web.filter.OncePerRequestFilter
+import javax.servlet.FilterChain
+import javax.servlet.http.HttpServletRequest
+import javax.servlet.http.HttpServletResponse
+
+
+@Component
+class JwtReqFilter(
+    val tokenProvider: TokenProvider,
+): OncePerRequestFilter() {
+    override fun doFilterInternal(request: HttpServletRequest, response: HttpServletResponse, filterChain: FilterChain) {
+        val accessToken = request.getHeader("Authorization")
+        if(accessToken!=null){
+            if (accessToken != null && tokenProvider.validateToken(accessToken)) {
+                val authentication: Authentication = tokenProvider.getAuthentication(accessToken)
+                SecurityContextHolder.getContext().authentication = authentication
+            }
+        }
+        filterChain.doFilter(request, response)
+    }
+}
