+++ src/main/kotlin/com/dotori/v2/global/error/ErrorResponse.kt
@@ -1,0 +1,6 @@
+package com.dotori.v2.global.error
+
+class ErrorResponse(
+    val message: String,
+    val code: Int,
+)
