--- src/main/kotlin/com/dotori/v2/infraStructure/config/aws/AwsSesConfig.kt
@@ -1,27 +1,0 @@
-package com.dotori.v2.infraStructure.config.aws
-
-import com.amazonaws.auth.AWSStaticCredentialsProvider
-import com.amazonaws.auth.BasicAWSCredentials
-import com.amazonaws.services.simpleemail.AmazonSimpleEmailService
-import com.amazonaws.services.simpleemail.AmazonSimpleEmailServiceClientBuilder
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-class AwsSesConfig {
-    @Value("\${aws.ses.credentials.access-key}")
-    private val accessKey: String? = null
-
-    @Value("\${aws.ses.credentials.secret-key}")
-    private val secretKey: String? = null
-    @Bean
-    fun amazonSimpleEmailService(): AmazonSimpleEmailService {
-        val basicAWSCredentials = BasicAWSCredentials(accessKey, secretKey)
-        val awsStaticCredentialsProvider = AWSStaticCredentialsProvider(basicAWSCredentials)
-        return AmazonSimpleEmailServiceClientBuilder.standard()
-            .withCredentials(awsStaticCredentialsProvider)
-            .withRegion("ap-northeast-2")
-            .build()
-    }
-}
