--- src/main/kotlin/com/dotori/v2/infraStructure/config/security/handler/CustomAuthenticationEntryPointHandler.kt
@@ -1,21 +1,0 @@
-package com.dotori.v2.infraStructure.config.security.handler
-
-import org.slf4j.LoggerFactory
-import org.springframework.security.core.AuthenticationException
-import org.springframework.security.web.AuthenticationEntryPoint
-import org.springframework.stereotype.Component
-import java.io.IOException
-import javax.servlet.ServletException
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
-
-@Component
-class CustomAuthenticationEntryPointHandler : AuthenticationEntryPoint {
-    private val log = LoggerFactory.getLogger(this::class.simpleName)
-
-    @Throws(IOException::class, ServletException::class)
-    override fun commence(request: HttpServletRequest, response: HttpServletResponse, authException: AuthenticationException) {
-        log.info("=== AuthenticationEntryPoint ===")
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED)
-    }
-}
