--- src/main/kotlin/com/dotori/v2/infraStructure/exception/ErrorResponse.kt
@@ -1,6 +1,0 @@
-package com.dotori.v2.infraStructure.exception
-
-class ErrorResponse(
-    val message: String,
-    val code: Int,
-)
