--- src/main/kotlin/com/dotori/v2/infraStructure/global/annotation/ReadOnlyUseCase.kt
@@ -1,8 +1,0 @@
-package com.dotori.v2.infraStructure.global.annotation
-
-import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
-
-@Service
-@Transactional(rollbackFor = [Exception::class], readOnly = true)
-annotation class ReadOnlyUseCase()
